requirements.txt
opencv-python
Pillow
face_recognition